場の複雑さ計算関数を追加,sticky_algoがそれに対応

+++ tengu/algorithm/sticky_algo.hpp
@@ -3,7 +3,7 @@
 #include <takao.hpp>
 struct evalated_field{
     field_type field;
-    int score;
+    double score;
 };
 struct putted_evalated_field{
     int flip;
@@ -11,7 +11,7 @@ struct putted_evalated_field{
     int y;
     int x;
     evalated_field e_field;
-    int score;
+    double score;
 };
 
 class sticky_algo : public algorithm_type
@@ -25,7 +25,7 @@ public:
 private:
     const problem_type origin_problem;
     problem_type problem;
-    int eval(field_type &field, const stone_type &stone, int pos_y, int pos_x);
+    double eval(field_type &field, const stone_type &stone, int pos_y, int pos_x);
     std::vector<evalated_field> eval_pattern(stone_type stone, std::vector<evalated_field> pattern, int search_width);
     std::vector<putted_evalated_field> result_stone;
 };
